Authorization in posts

+++ backend/routes/posts.js
@@ -27,12 +27,14 @@ const storage = multer.diskStorage({
   }
 });
 
+// Create post
 router.post("/", checkAuth, multer({storage: storage}).single("image"), (req, res, next) => {
   const url = req.protocol + '://' + req.get("host");
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath: url + "/images/" + req.file.filename
+    imagePath: url + "/images/" + req.file.filename,
+    creator: req.userData.userId
   });
   post.save().then(created => {
     res.status(201).json({
@@ -47,6 +49,7 @@ router.post("/", checkAuth, multer({storage: storage}).single("image"), (req, re
   });
 });
 
+// Get all posts
 router.get("/", (req, res, next) => {
   const pageSize = +req.query.pagesize;
   const currentPage = +req.query.page;
@@ -67,6 +70,7 @@ router.get("/", (req, res, next) => {
   });
 });
 
+// Get post by id
 router.get("/:id", (req, res, next) => {
   Post.findById(req.params.id).then((post) => {
     if (post) {
@@ -77,6 +81,7 @@ router.get("/:id", (req, res, next) => {
   });
 });
 
+// Update post by id
 router.put("/:id", checkAuth, multer({storage: storage}).single("image"), (req, res, next) => {
   let imagePath = req.body.imagePath;
   if (req.file) {
@@ -87,18 +92,26 @@ router.put("/:id", checkAuth, multer({storage: storage}).single("image"), (req, 
     _id: req.params.id,
     title: req.body.title,
     content: req.body.content,
-    imagePath: imagePath
+    imagePath: imagePath,
+    creator: req.userData.userId
   });
-  Post.updateOne({_id: req.params.id}, post).then((result) => {
-    console.log(result);
-    res.status(200).json({message: "Update successful!"});
+  Post.updateOne({_id: req.params.id, creator: req.userData.userId}, post).then((result) => {
+    if (result.nModified > 0) {
+      res.status(200).json({message: "Update successful!"});
+    } else {
+      res.status(401).json({message: "not authorized!"});
+    }
   })
 });
 
+// Delete post by id
 router.delete("/:id", checkAuth, (req, res, next) => {
-  Post.deleteOne({_id: req.params.id}).then((result) => {
-    console.log(result);
-    res.status(200).json({message: "Post deleted!"});
+  Post.deleteOne({_id: req.params.id, creator: req.userData.userId}).then((result) => {
+    if (result.n > 0) {
+      res.status(200).json({message: "Deletion successful!"});
+    } else {
+      res.status(401).json({message: "not authorized!"});
+    }
   })
 });
